80% done mong vay

+++ models/room.js
@@ -1,5 +1,5 @@
 "use strict";
-const { Model } = require("sequelize");
+const { Model, where } = require("sequelize");
 module.exports = (sequelize, DataTypes) => {
   class Room extends Model {
     /**
@@ -7,10 +7,11 @@ module.exports = (sequelize, DataTypes) => {
      * This method is not a part of Sequelize lifecycle.
      * The `models/index` file will call this method automatically.
      */
-    static associate({ Hotels, roomService, UrlImageRoom }) {
+    static associate({ Hotels, roomService, UrlImageRoom, Booking }) {
       this.belongsTo(Hotels, { foreignKey: "hotelId" });
       this.hasMany(roomService, { foreignKey: "roomId", onDelete: "CASCADE" });
       this.hasMany(UrlImageRoom, { foreignKey: "IdRoom", onDelete: "CASCADE" });
+      this.hasMany(Booking, { foreignKey: "room_id", onDelete: "CASCADE" });
     }
   }
   Room.init(
@@ -31,10 +32,19 @@ module.exports = (sequelize, DataTypes) => {
           const UrlImageRoom = sequelize.models.UrlImageRoom;
 
           // Xóa tất cả các bản ghi trong bảng UrlImageHotel có HotelId tương ứng
-          await UrlImageRoom.destroy({ where: { IdRoom: roomId } });
+          if (UrlImageRoom) {
+            await UrlImageRoom.destroy({ where: { IdRoom: roomId } });
+          }
 
           const roomService = sequelize.models.roomService;
-          await roomService.destroy({ where: { roomId: roomId } });
+          if (roomService) {
+            await roomService.destroy({ where: { roomId: roomId } });
+          }
+
+          const booking = sequelize.models.Booking;
+          if (booking) {
+            await booking.destroy({ where: { room_id: roomId } });
+          }
         },
       },
     }
